refactor: 쿠폰 utils 파일 분리 - 검증, 개별 계산, 통합계산, 최적화
