+++ src/features/coupon/utils/index.ts
@@ -1,0 +1,9 @@
+export { isCouponApplicable } from './couponValidation';
+export {
+  calculateFixedDiscount,
+  calculateBuyXGetYDiscount,
+  calculateShippingDiscount,
+  calculatePercentageDiscount,
+} from './couponDiscountCalculators';
+export { calculateCouponDiscount } from './singleCouponCalculator';
+export { calculateOptimalCouponDiscount } from './optimalCouponCalculator';
